backend/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-a2a>=1.0.0b260123",
    "agent-framework-ag-ui>=1.0.0b260123",
    "agent-framework-azure-ai>=1.0.0b260123",
    "agent-framework-core>=1.0.0b260123",
    "azure-ai-projects>=2.0.0b3",
    "azure-identity>=1.15.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.11.3",
    "python-dotenv",
    "fastapi-azure-auth>=5.0.0",
    "azure-monitor-opentelemetry>=1.6.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.0",
    "opentelemetry-exporter-otlp>=1.39.0",
]
